Remove unused code and general file cleanup build_Creator3.0.154

+++ CreatorCode/CreatorClient/Entities/ModuleProperty.cs
@@ -78,7 +78,7 @@ namespace Client.Creator
         {
             get
             {
-                return (uint)_commLink.GetMaxModuleValue(ProductID, (uint)ID);
+                return (uint)_commLink.GetMaxModuleValue(ProductID, ID);
             }
         }
